# net: replace RegExpPrototypeTest with RegExpPrototypeExec in isIP utils