--- a/lib/internal/net.js
+++ b/lib/internal/net.js
@@ -35,10 +35,7 @@ const IPv6Reg = new RegExp('^(?:' +
  * @returns {boolean}
  */
 function isIPv4(s) {
-  // TODO(aduh95): Replace RegExpPrototypeTest with RegExpPrototypeExec when it
-  // no longer creates a perf regression in the dns benchmark.
-  // eslint-disable-next-line node-core/avoid-prototype-pollution
-  return RegExpPrototypeTest(IPv4Reg, s);
+  return RegExpPrototypeExec(IPv4Reg, s) !== null;
 }
 
 /**
@@ -46,10 +43,7 @@ function isIPv4(s) {
  * @returns {boolean}
  */
 function isIPv6(s) {
-  // TODO(aduh95): Replace RegExpPrototypeTest with RegExpPrototypeExec when it
-  // no longer creates a perf regression in the dns benchmark.
-  // eslint-disable-next-line node-core/avoid-prototype-pollution
-  return RegExpPrototypeTest(IPv6Reg, s);
+  return RegExpPrototypeExec(IPv6Reg, s) !== null;
 }
 
 /**