--- a/lib/internal/net.js
+++ b/lib/internal/net.js
@@ -2,7 +2,7 @@
 
 const {
   RegExp,
-  RegExpPrototypeTest,
+  RegExpPrototypeExec,
   Symbol,
 } = primordials;