This commit replaces `RegExpPrototypeTest` with `RegExpPrototypeExec` in
`isIPv4` and `isIPv6` functions in `lib/internal/net.js`.

Originally, `RegExpPrototypeTest` was introduced due to a performance
regression observed in the `dns` benchmark (see TODO by @aduh95).
Re-running the benchmarks on current main shows that this regression
no longer exists.

### Benchmark results (macOS 15.0, Apple M1 Pro, Node.js v23.0.0-pre)

#### dns/lookup-promises.js

| Case | Before (ns/op) | After (ns/op) | Diff | % Change |
|------|----------------|---------------|------|-----------|
| 127.0.0.1, all=true | 12,359,656,105 | 15,929,908,403 | +3,570,252,298 | +28.9% |
| 127.0.0.1, all=false | 16,214,025,132 | 16,342,058,903 | +128,033,771 | +0.79% |
| ::1, all=true | 1,843,912,783 | 1,868,867,545 | +24,954,762 | +1.35% |
| ::1, all=false | 1,855,287,570 | 1,801,449,951 | -53,837,619 | -2.90% |

#### dns/lookup.js

| Case | Before (ns/op) | After (ns/op) | Diff | % Change |
|------|----------------|---------------|------|-----------|
| 127.0.0.1, all=true | 9,487,938 | 9,292,160 | -195,778 | -2.06% |
| 127.0.0.1, all=false | 10,361,554 | 10,604,750 | +243,196 | +2.35% |
| ::1, all=true | 6,576,433 | 6,696,469 | +120,036 | +1.83% |
| ::1, all=false | 7,040,495 | 7,133,674 | +93,179 | +1.32% |

### Analysis

- `dns/lookup-promises.js`: IPv4 all=true shows the largest improvement (+28.9%)  
- IPv6 some cases show slight decrease (-2.9%), but overall stable  
- `dns/lookup.js`: mostly small improvements (+2%), minor decreases (-2%)  
- Overall, applying `RegExpPrototypeExec` does **not introduce regressions** and shows **stable performance improvements**.

<!--
Before submitting a pull request, please read:

- the CONTRIBUTING guide at https://github.com/nodejs/node/blob/HEAD/CONTRIBUTING.md
- the commit message formatting guidelines at
  https://github.com/nodejs/node/blob/HEAD/doc/contributing/pull-requests.md#commit-message-guidelines

For code changes:
1. Include tests for any bug fixes or new features.
2. Update documentation if relevant.
3. Ensure that `make -j4 test` (UNIX), or `vcbuild test` (Windows) passes.

If you believe this PR should be highlighted in the Node.js CHANGELOG
please add the `notable-change` label.

Developer's Certificate of Origin 1.1

By making a contribution to this project, I certify that:

(a) The contribution was created in whole or in part by me and I
    have the right to submit it under the open source license
    indicated in the file; or

(b) The contribution is based upon previous work that, to the best
    of my knowledge, is covered under an appropriate open source
    license and I have the right under that license to submit that
    work with modifications, whether created in whole or in part
    by me, under the same open source license (unless I am
    permitted to submit under a different license), as indicated
    in the file; or

(c) The contribution was provided directly to me by some other
    person who certified (a), (b) or (c) and I have not modified
    it.

(d) I understand and agree that this project and the contribution
    are public and that a record of the contribution (including all
    personal information I submit with it, including my sign-off) is
    maintained indefinitely and may be redistributed consistent with
    this project or the open source license(s) involved.
-->