@haramj 2025-08-25
I noticed that on my local machine (macOS M1 Pro) the benchmarks showed improvements (up to +28.9% in some cases),
but the CI benchmark results indicate regressions (up to ~17%).

Would you recommend additional cases or environments I should benchmark to get a clearer picture?
I’d be happy to re-run tests or provide more detailed comparisons.