@aduh95 2025-08-28
I've tried running the benchmark on GHA in https://github.com/aduh95/node/actions/runs/17297577621/job/49099592376 (using Apple M1 chip), and it's showing similar results:

```
                                                              confidence improvement accuracy (*)    (**)   (***)
dns/lookup-promises.js n=5000000 all='false' name='::1'                      -1.82 %       ±7.95% ±10.58% ±13.78%
dns/lookup-promises.js n=5000000 all='false' name='127.0.0.1'          *    -12.62 %      ±12.42% ±16.55% ±21.59%
dns/lookup-promises.js n=5000000 all='true' name='::1'                       -1.23 %       ±8.85% ±11.79% ±15.38%
dns/lookup-promises.js n=5000000 all='true' name='127.0.0.1'                  2.07 %       ±9.87% ±13.15% ±17.15%
dns/lookup.js n=5000000 all='false' name='::1'                         *     -6.07 %       ±5.85%  ±7.79% ±10.14%
dns/lookup.js n=5000000 all='false' name='127.0.0.1'                 ***    -12.13 %       ±5.50%  ±7.32%  ±9.53%
dns/lookup.js n=5000000 all='true' name='::1'                                -4.83 %       ±5.19%  ±6.91%  ±9.00%
dns/lookup.js n=5000000 all='true' name='127.0.0.1'                   **     -9.53 %       ±5.98%  ±7.96% ±10.37%
```

Maybe you were running the benchmark against the wrong binaries, or something in your configuration or the way you run the benchmarks makes you get irrelevant results.