@haramj 2025-08-29
Thanks for re-running the benchmarks and confirming the regression.  
I’ll go ahead and close this PR.